+++ volmdlr/faces.py
@@ -115,10 +115,8 @@ class Face3D(volmdlr.core.Primitive3D):
             f"overloaded by {self.__class__.__name__}")
 
     def get_bounding_box(self):
-        """Abstract method."""
-        raise NotImplementedError(
-            f"self.__class__.__name__"
-            f"overloaded by {self.__class__.__name__}")
+        """General method to get the bounding box of a face 3D."""
+        return self.outer_contour3d.bounding_box
 
     def area(self):
         return self.surface2d.area()
@@ -293,6 +291,8 @@ class Face3D(volmdlr.core.Primitive3D):
     def triangulation(self):
         number_points_x, number_points_y = self.grid_size()
         mesh2d = self.surface2d.triangulation(number_points_x, number_points_y)
+        if mesh2d is None:
+            return None
         return vmd.DisplayMesh3D([vmd.Node3D(*self.surface3d.point2d_to_3d(point)) for point in mesh2d.points],
                                  mesh2d.triangles)
 
@@ -431,6 +431,8 @@ class Face3D(volmdlr.core.Primitive3D):
         linesegment_intersections = []
         if not self.bounding_box.bbox_intersection(linesegment.bounding_box):
             return []
+        if not hasattr(self.surface3d, 'linesegment_intersections'):
+            return self.linesegment_intersections_approximation(linesegment)
         for intersection in self.surface3d.linesegment_intersections(linesegment):
             if self.point_belongs(intersection):
                 linesegment_intersections.append(intersection)
@@ -1120,7 +1122,10 @@ class Face3D(volmdlr.core.Primitive3D):
 
     def is_linesegment_crossing(self, linesegment):
         """Verify if a face 3d is being crossed by a linesegment 3d. """
-        if not self.bounding_box.bbox_intersection(linesegment.bounding_box):
+        if math.isclose(self.area(), 0.0, abs_tol=1e-10):
+            return False
+        bbox_block_faces = volmdlr.primitives3d.Block.from_bounding_box(self.bounding_box).faces
+        if not any(bbox_face.linesegment_intersections(linesegment) for bbox_face in bbox_block_faces):
             return False
         triangulation = self.triangulation()
         faces_triangulation = triangulation.triangular_faces()
@@ -1129,6 +1134,23 @@ class Face3D(volmdlr.core.Primitive3D):
             if inters:
                 return True
         return False
+
+    def linesegment_intersections_approximation(self, linesegment: vme.LineSegment3D) -> List[volmdlr.Point3D]:
+        """Approximation of intersections between a b-spline face 3D and a line segment 3D."""
+        if math.isclose(self.area(), 0.0, abs_tol=1e-10):
+            return False
+        bbox_block_faces = volmdlr.primitives3d.Block.from_bounding_box(self.bounding_box).faces
+        if not any(bbox_face.linesegment_intersections(linesegment) for bbox_face in bbox_block_faces):
+            return []
+        triangulation = self.triangulation()
+        faces_triangulation = triangulation.triangular_faces()
+        linesegment_intersections = []
+        for face in faces_triangulation:
+            inters = face.linesegment_intersections(linesegment)
+            for point in inters:
+                if not volmdlr.core.point_in_list(point, linesegment_intersections):
+                    linesegment_intersections.append(point)
+        return linesegment_intersections
 
 
 class PlaneFace3D(Face3D):
@@ -1172,9 +1194,6 @@ class PlaneFace3D(Face3D):
     @bounding_box.setter
     def bounding_box(self, new_bounding_box):
         self._bbox = new_bounding_box
-
-    def get_bounding_box(self):
-        return self.outer_contour3d.bounding_box
 
     def distance_to_point(self, point, return_other_point=False):
         """
@@ -1948,14 +1967,8 @@ class CylindricalFace3D(Face3D):
     def bounding_box(self, new_bouding_box):
         self._bbox = new_bouding_box
 
-    def get_bounding_box(self):
-        """
-        Computes the bounding box using the contour 3d. true in this case of cylindrical face (not general).
-        """
-        return self.outer_contour3d.bounding_box
-
     def triangulation_lines(self, angle_resolution=5):
-        theta_min, theta_max, zmin, zmax = self.surface2d.bounding_rectangle.bounds()
+        theta_min, theta_max, zmin, zmax = self.surface2d.bounding_rectangle().bounds()
         delta_theta = theta_max - theta_min
         nlines = math.ceil(delta_theta * angle_resolution)
         lines = []
@@ -2168,9 +2181,6 @@ class ToroidalFace3D(Face3D):
     def bounding_box(self, new_bounding_box):
         self._bbox = new_bounding_box
 
-    def get_bounding_box(self):
-        return self.surface3d.bounding_box
-
     def triangulation_lines(self, angle_resolution=5):
         theta_min, theta_max, phi_min, phi_max = self.surface2d.bounding_rectangle().bounds()
 
@@ -2274,51 +2284,6 @@ class ConicalFace3D(Face3D):
     @bounding_box.setter
     def bounding_box(self, new_bouding_box):
         self._bbox = new_bouding_box
-
-    def get_bounding_box(self):
-        theta_min, theta_max, zmin, zmax = self.surface2d.outer_contour.bounding_rectangle.bounds()
-
-        x_vector = (volmdlr.X3D.dot(self.surface3d.frame.u) * self.surface3d.frame.u
-                    + volmdlr.X3D.dot(self.surface3d.frame.v) * self.surface3d.frame.v)
-        try:
-            x_vector.normalize()
-        except ZeroDivisionError:
-            pass
-        y_vector = (volmdlr.Y3D.dot(self.surface3d.frame.u) * self.surface3d.frame.u
-                    + volmdlr.Y3D.dot(self.surface3d.frame.v) * self.surface3d.frame.v)
-
-        try:
-            y_vector.normalize()
-        except ZeroDivisionError:
-            pass
-
-        z_vector = (volmdlr.Z3D.dot(self.surface3d.frame.u) * self.surface3d.frame.u
-                    + volmdlr.Z3D.dot(self.surface3d.frame.v) * self.surface3d.frame.v)
-        try:
-            z_vector.normalize()
-        except ZeroDivisionError:
-            pass
-
-        lower_center = self.surface3d.frame.origin + zmin * self.surface3d.frame.w
-        upper_center = self.surface3d.frame.origin + zmax * self.surface3d.frame.w
-        lower_radius = math.tan(self.surface3d.semi_angle) * zmin
-        upper_radius = math.tan(self.surface3d.semi_angle) * zmax
-
-        points = [lower_center - lower_radius * x_vector,
-                  lower_center + lower_radius * x_vector,
-                  lower_center - lower_radius * y_vector,
-                  lower_center + lower_radius * y_vector,
-                  lower_center - lower_radius * z_vector,
-                  lower_center + lower_radius * z_vector,
-                  upper_center - upper_radius * x_vector,
-                  upper_center + upper_radius * x_vector,
-                  upper_center - upper_radius * y_vector,
-                  upper_center + upper_radius * y_vector,
-                  upper_center - upper_radius * z_vector,
-                  upper_center + upper_radius * z_vector,
-                  ]
-
-        return volmdlr.core.BoundingBox.from_points(points)
 
     def triangulation_lines(self, angle_resolution=5):
         theta_min, theta_max, zmin, zmax = self.surface2d.bounding_rectangle().bounds()
@@ -2430,10 +2395,6 @@ class SphericalFace3D(Face3D):
     @bounding_box.setter
     def bounding_box(self, new_bouding_box):
         self._bbox = new_bouding_box
-
-    def get_bounding_box(self):
-        # To be enhanced
-        return self.surface3d.bounding_box
 
     def triangulation_lines(self, angle_resolution=7):
         """
@@ -2630,12 +2591,6 @@ class ExtrusionFace3D(Face3D):
     def bounding_box(self, new_bouding_box):
         self._bbox = new_bouding_box
 
-    def get_bounding_box(self):
-        # To be enhanced by restricting wires to cut
-        points = self.outer_contour3d.discretization_points(number_points=25)
-
-        return volmdlr.core.BoundingBox.from_points(points)
-
     def grid_size(self):
         """
         Specifies an adapted size of the discretization grid used in face triangulation.
@@ -2703,21 +2658,6 @@ class RevolutionFace3D(Face3D):
     def bounding_box(self, new_bouding_box):
         self._bbox = new_bouding_box
 
-    def get_bounding_box(self):
-        # To be enhanced by restricting wires to cut
-        try:
-            points = self.outer_contour3d.discretization_points(angle_resolution=20)
-        except Exception:
-            self.outer_contour3d.plot()
-            self.outer_contour3d
-        # points = []
-        # for i in range(37):
-        #     angle = i * volmdlr.TWO_PI / 36
-        #     points.extend([point.rotation(self.surface3d.axis_point, self.surface3d.axis, angle)
-        #                    for point in curve_points])
-
-        return volmdlr.core.BoundingBox.from_points(points)
-
     def grid_size(self):
         """
         Specifies an adapted size of the discretization grid used in face triangulation.
@@ -2781,9 +2721,6 @@ class BSplineFace3D(Face3D):
     @bounding_box.setter
     def bounding_box(self, new_bounding_box):
         self._bbox = new_bounding_box
-
-    def get_bounding_box(self):
-        return self.surface3d.bounding_box
 
     def triangulation_lines(self, resolution=25):
         u_min, u_max, v_min, v_max = self.surface2d.bounding_rectangle().bounds()
@@ -3157,15 +3094,3 @@ class BSplineFace3D(Face3D):
             inner_contours=[plane3d.contour3d_to_2d(contour) for contour in self.inner_contours3d])
 
         return PlaneFace3D(surface3d=plane3d, surface2d=surface2d)
-
-    def linesegment_intersections(self, linesegment: vme.LineSegment3D) -> List[volmdlr.Point3D]:
-        """Aproximation of intersections between a bspline face 3D and a line segment 3D."""
-        triangulation = self.triangulation()
-        faces_triangulation = triangulation.triangular_faces()
-        linesegment_intersections = []
-        for face in faces_triangulation:
-            inters = face.linesegment_intersections(linesegment)
-            for point in inters:
-                if volmdlr.core.point_in_list(point, linesegment_intersections):
-                    linesegment_intersections.append(point)
-        return linesegment_intersections